backend/requirements.txt

Flask
python-dotenv
groq
Flask-Cors
